Peek inside the right file
At least I think this is what it's supposed to say, given that we only
inspect the executable instead of the object file.
Because the `a.out` is kind of a random name, maybe rewriting it to
"and peek inside the result" or "peek inside the executable" would be
the better phrasing.