driver.php: updated gcc -v example
The original example omitted the important `cc1` step, which was probably an oversight. It also had different temporary filenames between the executions, also an oversight. It also had `==` instead of just `=`, which was a simple typo.

This commit here tries to make the output more faithful (all three steps are visible), while redacting a lot of extra fluff from the output. In fact, we can redact even more:

    gcc -v hello.c 2>&1 | grep '^ [^ ].* '

That will output just the lines starting with a single space, and that contain at least another space somewhere. That's enough to extract the three execution lines from the noisy output. They still have countless parameters in them, which need to be cleaned up manually.